chore: add refer friend and invite code login screen
