+++ app/src/main/java/com/getcode/view/main/getKin/GetKinSheet.kt
@@ -1,74 +1,236 @@
 package com.getcode.view.main.getKin
 
-import androidx.compose.foundation.layout.fillMaxHeight
+import androidx.compose.animation.AnimatedVisibility
+import androidx.compose.foundation.Image
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.fillMaxWidth
-import androidx.compose.runtime.*
+import androidx.compose.foundation.layout.height
+import androidx.compose.foundation.layout.padding
+import androidx.compose.foundation.layout.size
+import androidx.compose.material.CircularProgressIndicator
+import androidx.compose.material.MaterialTheme
+import androidx.compose.material.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.collectAsState
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
-import androidx.compose.ui.platform.LocalLifecycleOwner
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.res.colorResource
+import androidx.compose.ui.res.painterResource
 import androidx.compose.ui.res.stringResource
-import androidx.lifecycle.Lifecycle
-import androidx.navigation.NavOptions
-import com.getcode.MainDestinations
-import com.getcode.rememberCodeAppState
-import com.getcode.theme.sheetHeight
-import com.getcode.view.SheetSections
-import com.getcode.view.components.SheetTitle
+import androidx.compose.ui.text.style.TextDecoration
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import androidx.constraintlayout.compose.ConstraintLayout
+import com.getcode.R
+import com.getcode.navigation.core.LocalCodeNavigator
+import com.getcode.navigation.screens.BuySellScreen
+import com.getcode.navigation.screens.ReferFriendScreen
+import com.getcode.theme.White
+import com.getcode.theme.White05
+import com.getcode.util.addIf
 import com.getcode.view.main.home.HomeViewModel
-import androidx.navigation.compose.NavHost
-import com.getcode.analytics.AnalyticsScreenWatcher
-import com.getcode.manager.AnalyticsManager
-import com.getcode.util.RepeatOnLifecycle
 
+data class GetKinItem(
+    val imageResId: Int,
+    val inactiveImageResId: Int,
+    val titleTextResId: Int,
+    val subtitleTextResId: Int? = null,
+    val isVisible: Boolean,
+    val isActive: Boolean,
+    val isLoading: Boolean,
+    val isStrikeThrough: Boolean,
+    val onClick: () -> Unit,
+)
 
 @Composable
 fun GetKinSheet(
-    homeViewModel: HomeViewModel,
-    onClose: () -> Unit = {}
+    viewModel: GetKinSheetViewModel,
+    homeViewModel: HomeViewModel
 ) {
-    var title: Int? by remember { mutableStateOf(null) }
-    var isBackButtonVisible: Boolean by remember { mutableStateOf(false) }
-    val appState = rememberCodeAppState()
+    val navigator = LocalCodeNavigator.current
+    val dataState by viewModel.uiFlow.collectAsState()
 
-    AnalyticsScreenWatcher(
-        lifecycleOwner = LocalLifecycleOwner.current,
-        event = AnalyticsManager.Screen.GetKin
+    val items = listOf(
+        GetKinItem(
+            imageResId = R.drawable.ic_gift,
+            inactiveImageResId = R.drawable.ic_gift_inactive,
+            titleTextResId = R.string.subtitle_getYourFirstKinFree,
+            subtitleTextResId = R.string.title_limitedTimeOffer,
+            isVisible = true,
+            isActive = dataState.isEligibleGetFirstKinAirdrop,
+            isLoading = dataState.isGetFirstKinAirdropLoading,
+            isStrikeThrough = !dataState.isEligibleGetFirstKinAirdrop,
+            onClick = {
+                if (!dataState.isEligibleGetFirstKinAirdrop || dataState.isGetFirstKinAirdropLoading) {
+                    return@GetKinItem
+                }
+
+                viewModel.requestFirstKinAirdrop({ navigator.hide() }, homeViewModel)
+            },
+        ),
+        GetKinItem(
+            imageResId = R.drawable.ic_send2,
+            inactiveImageResId = R.drawable.ic_send2_inactive,
+            titleTextResId = R.string.title_referFriend,
+            subtitleTextResId = R.string.title_limitedTimeOffer,
+            isVisible = dataState.isEligibleGiveFirstKinAirdrop,
+            isActive = dataState.isEligibleGiveFirstKinAirdrop,
+            isLoading = false,
+            isStrikeThrough = !dataState.isEligibleGiveFirstKinAirdrop,
+            onClick = {
+                if (!dataState.isEligibleGiveFirstKinAirdrop) {
+                    return@GetKinItem
+                }
+
+                navigator.push(ReferFriendScreen)
+//                navigator.push()
+//                navController.navigate(SheetSections.REFER_FRIEND.route)
+            },
+        ),
+        GetKinItem(
+            imageResId = R.drawable.ic_currency_dollar_active,
+            inactiveImageResId = R.drawable.ic_currency_dollar_inactive,
+            titleTextResId = R.string.subtitle_buyAndSendKin,
+            isVisible = true,
+            isActive = true,
+            isLoading = false,
+            isStrikeThrough = false,
+            onClick = {
+                navigator.push(BuySellScreen)
+            },
+        ),
     )
 
-    SheetTitle(
-        title = title?.let { stringResource(id = it) },
-        backButton = isBackButtonVisible,
-        closeButton = !isBackButtonVisible,
-        onBackIconClicked = { appState.sheetNavController.popBackStack() },
-        onCloseIconClicked = onClose
-    )
-
-    NavHost(
+    ConstraintLayout(
         modifier = Modifier
             .fillMaxWidth()
-            .fillMaxHeight(sheetHeight),
-        navController = appState.sheetNavController,
-        startDestination = MainDestinations.SHEET_GRAPH,
+            .padding(horizontal = 20.dp),
     ) {
-//        codeSheetNavGraph(
-//            navController = appState.sheetNavController,
-//            homeViewModel = homeViewModel,
-//            onTitleChange = { title = it },
-//            onBackButtonVisibilityChange = { isBackButtonVisible = it },
-//            onClose = onClose
-//        )
-    }
+        val (topSection, bottomSection) = createRefs()
 
-    RepeatOnLifecycle(targetState = Lifecycle.State.RESUMED) {
-        val isHome =
-            appState.sheetNavController.currentDestination?.route == SheetSections.GET_KIN.route
-        if (!isHome) {
-            appState.sheetNavController.popBackStack()
-            appState.sheetNavController.navigate(
-                SheetSections.GET_KIN.route,
-                NavOptions.Builder().setPopUpTo(
-                    SheetSections.NONE.route, inclusive = true, saveState = false
-                ).build()
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .constrainAs(topSection) {
+                    top.linkTo(parent.top)
+                    start.linkTo(parent.start)
+                    end.linkTo(parent.end)
+                },
+        ) {
+            Image(
+                painter = painterResource(R.drawable.ic_graphic_wallet),
+                contentDescription = "",
+                modifier = Modifier.padding(vertical = 10.dp),
             )
+            Text(
+                text = stringResource(R.string.title_getKin),
+                style = MaterialTheme.typography.h1,
+                modifier = Modifier.padding(vertical = 15.dp),
+            )
+            Text(
+                text = stringResource(R.string.subtitle_getKin),
+                style = MaterialTheme.typography.body1,
+                modifier = Modifier.padding(vertical = 10.dp),
+            )
+        }
+
+        Column(
+            modifier = Modifier
+                .fillMaxWidth()
+                .constrainAs(bottomSection) {
+                    top.linkTo(topSection.bottom)
+                    start.linkTo(parent.start)
+                    end.linkTo(parent.end)
+                    bottom.linkTo(parent.bottom)
+                },
+        ) {
+            Column {
+                for (item in items) {
+                    if (!item.isVisible) {
+                        continue
+                    }
+
+                    Spacer(
+                        modifier = Modifier
+                            .fillMaxWidth()
+                            .height(1.dp)
+                            .background(White05),
+                    )
+
+                    Row(
+                        modifier = Modifier
+                            .addIf(
+                                item.isStrikeThrough.not(),
+                            ) {
+                                Modifier.clickable { item.onClick() }
+                            }
+                            .padding(vertical = 20.dp, horizontal = 10.dp),
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        Image(
+                            modifier = Modifier.size(25.dp),
+                            painter = if (item.isActive) painterResource(id = item.imageResId) else painterResource(id = item.inactiveImageResId),
+                            contentDescription = "",
+                        )
+                        Column(
+                            modifier = Modifier
+                                .padding(start = 17.dp)
+                                .weight(1f),
+                        ) {
+                            Text(
+                                text = stringResource(item.titleTextResId),
+                                color = if (item.isActive) Color.White else colorResource(R.color.code_brand_light),
+                                style = MaterialTheme.typography.button.copy(
+                                    textDecoration = if (item.isStrikeThrough) TextDecoration.LineThrough else MaterialTheme.typography.button.textDecoration,
+                                ),
+                            )
+                            item.subtitleTextResId?.let {
+                                Text(
+                                    modifier = Modifier.padding(top = 3.dp),
+                                    text = stringResource(it),
+                                    style = MaterialTheme.typography.caption,
+                                    fontSize = 13.sp,
+                                    color = colorResource(R.color.code_brand_light),
+                                )
+                            }
+                        }
+
+                        if (item.isLoading) {
+                            CircularProgressIndicator(
+                                strokeWidth = 2.dp,
+                                color = White,
+                                modifier = Modifier
+                                    .size(20.dp)
+                                    .align(Alignment.CenterVertically),
+                            )
+                        } else if (item.isActive) {
+                            Image(
+                                modifier = Modifier.padding(start = 10.dp),
+                                painter = painterResource(R.drawable.ic_chevron_right),
+                                contentDescription = "",
+                            )
+                        }
+                    }
+                }
+            }
+
+            Spacer(
+                modifier = Modifier
+                    .fillMaxWidth()
+                    .height(1.dp)
+                    .background(White05),
+            )
+
+            // TODO: need a better way to reload model
+            AnimatedVisibility(visible = true) {
+                viewModel.reset()
+            }
         }
     }
 }
